+++ src/graphics/configSimulation.hpp
@@ -1,25 +1,25 @@
 #ifndef CONFIGSIMULATION_HPP
 #define CONFIGSIMULATION_HPP
-#include "../simulation/System.hpp"
-#include "SimulationWindow.hpp"
+#include "../simulation/system.hpp"
+#include "simulationWindow.hpp"
 
-#include "AppStyle.hpp"
-#include "Scene.hpp"
-#include "components/button/Button.hpp"
-#include "components/textInput/TextInput.hpp"
+#include "appStyle.hpp"
+#include "components/button/button.hpp"
+#include "components/textInput/textInput.hpp"
+#include "scene.hpp"
 #include <SFML/Graphics.hpp>
 #include <memory>
 
 class ConfigSimulation {
   public:
     ConfigSimulation(std::shared_ptr<sf::RenderWindow> window,
-                     std::shared_ptr<System<double>> system,
+                     std::shared_ptr<System<float>> system,
                      Scene &selectedScene, SimulationWindow &simulationWindow);
     void draw();
     void processEvents();
 
   private:
-    std::shared_ptr<System<double>> system;
+    std::shared_ptr<System<float>> system;
     std::shared_ptr<sf::RenderWindow> window;
     Scene &selectedScene;
     SimulationWindow &simulationWindow;
