+++ src/graphics/mainWindow.hpp
@@ -1,12 +1,12 @@
 #ifndef MAINWINDOW_HPP
 #define MAINWINDOW_HPP
 
-#include "../simulation/System.hpp"
-#include "AnalysisWindow.hpp"
-#include "ConfigSimulation.hpp"
-#include "Menu.hpp"
-#include "Scene.hpp"
-#include "SimulationWindow.hpp"
+#include "../simulation/system.hpp"
+#include "analysisWindow.hpp"
+#include "configSimulation.hpp"
+#include "menu.hpp"
+#include "scene.hpp"
+#include "simulationWindow.hpp"
 #include <SFML/Graphics.hpp>
 #include <memory>
 
@@ -17,7 +17,7 @@ class MainWindow {
     Scene selectedScene;
 
   private:
-    std::shared_ptr<System<double>> system;
+    std::shared_ptr<System<float>> system;
     std::shared_ptr<sf::RenderWindow> window;
     Menu menu;
     ConfigSimulation configSimulation;
